move browsers to own file
